Feature/#186 마일스톤 권한 설정 (#187)
* refactor: jwt argument Resolver 명 변경

#186

* refactor: jwt annotation 명 수정

#186

* refactor: jwt annotation 명 수정

#186

* feat: student argument resolver 추가

#186

* feat: 마일스톤 등록 시, 해당 회원 권환 확인

#186

* feat: 학생만 마일스톤 삭제할 수 있도록 권한 설정

#186

* feat: 마일스톤 조회 권한 설정

#186

* feat: 마일스톤 점수 조회 api 불러오는 것 권한 설정

#186

* feat: 관리자의 마일스톤 권한 설정

#186

* feat: 프론트 오류 수정 및 api 에 authentication 추가

#186

* feat: api 호출 시 Authorization 붙여서 보내도록 수정

#186

* fix: page 폴더 바깥에서 next/header 사용할 수 없어서 직접 token 넣어주도록 변경

#186

+++ backend/src/main/java/sw_css/milestone/application/MilestoneHistoryCommandService.java
@@ -13,9 +13,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import sw_css.member.domain.StudentMember;
-import sw_css.member.domain.repository.StudentMemberRepository;
-import sw_css.member.exception.MemberException;
-import sw_css.member.exception.MemberExceptionType;
 import sw_css.milestone.application.dto.request.MilestoneHistoryCreateRequest;
 import sw_css.milestone.domain.Milestone;
 import sw_css.milestone.domain.MilestoneHistory;
@@ -35,19 +32,15 @@ public class MilestoneHistoryCommandService {
     private String filePathPrefix;
 
     // TODO 테스트 작성
-    private final StudentMemberRepository studentMemberRepository;
     private final MilestoneRepository milestoneRepository;
     private final MilestoneHistoryRepository milestoneHistoryRepository;
 
-    public Long registerMilestoneHistory(final MultipartFile file, final MilestoneHistoryCreateRequest request) {
+    public Long registerMilestoneHistory(final StudentMember student, final MultipartFile file,
+                                         final MilestoneHistoryCreateRequest request) {
         validateFileType(file);
 
         final Milestone milestone = milestoneRepository.findById(request.milestoneId())
                 .orElseThrow(() -> new MilestoneException(MilestoneExceptionType.NOT_FOUND_MILESTONE));
-        // TODO 요청자의 학번을 불러오는 로직 추가
-        final StudentMember student = studentMemberRepository.findById(202055558L).orElseThrow(
-                () -> new MemberException(MemberExceptionType.NOT_FOUND_STUDENT)
-        );
 
         final String newFilePath = generateFilePath(file);
         final MilestoneHistory newMilestoneHistory = new MilestoneHistory(milestone, student, request.description(),
@@ -100,11 +93,13 @@ public class MilestoneHistoryCommandService {
         }
     }
 
-    public void deleteMilestoneHistory(final Long historyId) {
+    public void deleteMilestoneHistory(final StudentMember student, final Long historyId) {
         final MilestoneHistory history = milestoneHistoryRepository.findById(historyId)
                 .orElseThrow(
                         () -> new MilestoneHistoryException(MilestoneHistoryExceptionType.NOT_FOUND_MILESTONE_HISTORY));
-
+        if (!history.getStudentId().equals(student.getId())) {
+            throw new MilestoneHistoryException(MilestoneHistoryExceptionType.REMOVE_NOT_ALLOWED);
+        }
         history.delete();
     }
 }
